+++ Changelog.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Update rustfmt's defaults to use 2021 edition ([#554](https://github.com/rust-lang/rust-mode/issues/509)).
+- Introduce `rust-format-mode` for `rust-format-buffer` ([#556](https://github.com/rust-lang/rust-mode/pull/556)).
 
 # v1.0.6